refactor(codegen): extract std::text::{byte_pred, base64::*} into stdlib/text.rs (Round 1.8 partial)
3 fns lifted into TextStdlib<'ctx>: byte_pred (multi-symbol dispatcher
covering is_alpha/is_digit/is_alnum/is_whitespace/is_word_char),
base64_encode, base64_decode.

std::text::tokenize_words_into stays in codegen.rs for now — it
reads LocusInfo / CapacitySlotLayout / SlotForm internals that
properly belong with Round 4 (locus extraction). Lifting it would
mean publishing locus-private surface ahead of when that round
designs the boundary cleanly.

codegen.rs: 43754 -> 43543 (-211 lines).
