+++ crates/hale-codegen/src/stdlib/mod.rs
@@ -15,4 +15,5 @@ pub(crate) mod env;
 pub(crate) mod math;
 pub(crate) mod process;
 pub(crate) mod rand;
+pub(crate) mod text;
 pub(crate) mod time;
